+++ src/definitions/list.ts
@@ -15,6 +15,21 @@ export const ASC = SortOrder.ASC;
 /** A descending sort. */
 export const DESC = SortOrder.DESC;
 
+/** A view mode for the list. */
+export interface ListMode {
+  /**
+   * The ID of the view mode.
+   * This is what is actually stored on the list state.
+   */
+  id: string;
+
+  /** The style of the view mode. */
+  style: string;
+
+  /** The label of the view mode. */
+  label: string;
+}
+
 /** The state of sorting a property on a list. */
 export interface SortState {
   /** The property that is being sorted. */
@@ -56,15 +71,45 @@ export interface PagingState {
   itemsPerPage: number;
 }
 
+/** The total state of the list. */
+export interface ListState {
+  /** The filter state of the list. */
+  filters: FilterState[];
+
+  /** The sorting state of the list. */
+  sorting: SortState[];
+
+  /** The pagination state of the list. */
+  paging?: PagingState;
+
+  /**
+   * Properties of the relevant model that have been
+   * selected to be visible. If the list definition
+   * has no custom `visible` function, this will default
+   * to all fields.
+   */
+  visible: Property<any>[];
+
+  /** The current list view mode. */
+  mode?: string;
+}
+
 /**
  * The definition of a list component.
  *
  * @param T The model
  */
 export interface ListDefinition<T extends Model> extends BaseDefinition<T> {
-  /** The callback for refreshing the list data from the current list state. */
-  refresh(filters: FilterState[], sorting: SortState[], paging: PagingState): Promise<T[]>;
+  /** The list view modes. */
+  modes: ListMode[];
 
-  /** Select the properties that should be visible on the list. */
-  visible(props: ModelProperties<T>): [Property<any>];
+  /** The callback for refreshing the list data from the current list state. */
+  refresh(state: ListState): Promise<T[]>;
+
+  /**
+   * Select the properties that should be visible on the list
+   * based off the list state. This is automatically called whenever
+   * the list state changes.
+   */
+  visible?(state: ListState, props: ModelProperties<T>): [Property<any>];
 }